+++ dhareshsawLescargot/src/app/character.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CharacterService } from './character.service';
+
+describe('CharacterService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: CharacterService = TestBed.get(CharacterService);
+    expect(service).toBeTruthy();
+  });
+});
